requirements.txt
asgiref==3.7.2
astroid==2.5.6
backcall==0.2.0
colorama==0.4.4
decorator==5.0.9
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.1.0
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.18
Pygments==2.9.0
pylint==2.8.2
python-dateutil==2.8.1
pywin32==300
pyzmq==22.0.3
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.5
wrapt==1.12.1
